refactor: combine savegame reader and linkgraph
This moves the existing savegame reader and linkgraph code into openttdlab.py.
This is an as close to copy and paste from their original files as possible,
while preserving the functionality of existing code in openttdlab.py

This is done to make it as friendly as possible to not only be used in
openttdlab.py, but be refactored and possibly made hidden from client code (as
much as Python makes that possible at least).

This deliberately includes a notice at the top of the file at the same time to
try to make clear who owns what copyright - what is mine and what is
TrueBrain's. Am opting to not include the year which is often deemed
unneecessary for files in source control these days. I hope this is still
abididing by GPL v2 which on modification+distribution:

> a) You must cause the modified files to carry prominent notices
>    stating that you changed the files and the date of any change.

The date is in source control - I hope this is sufficiently prominent.
